ODY-941 #close Handle NULL values in aggregation functions and update testcases
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@24312 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 633843cb6fb65f266540681e682beda00d726a69

+++ server/relational/relational.base/src/de/uniol/inf/is/odysseus/physicaloperator/relational/RelationalAvgSum.java
@@ -45,7 +45,7 @@ public class RelationalAvgSum extends AvgSum<Tuple<?>, Tuple<?>> {
 			return init((AvgSumPartialAggregate<Tuple<?>>) in.getAttribute(pos));
 		} else {
 			return new AvgSumPartialAggregate<Tuple<?>>(
-					((Number) in.getAttribute(pos)).doubleValue(), 1);
+					((Number) in.getAttribute(pos)), 1);
 		}
 	}
 
@@ -57,8 +57,7 @@ public class RelationalAvgSum extends AvgSum<Tuple<?>, Tuple<?>> {
 			return merge(p, (IPartialAggregate) toMerge.getAttribute(pos),
 					false);
 		} else {
-			return pa.addAggValue(((Number) toMerge.getAttribute(pos))
-					.doubleValue());
+			return pa.addAggValue(((Number) toMerge.getAttribute(pos)));
 		}
 	}
